requirements.txt
fastapi
uvicorn
requests
fastapi-cache2
pandas
yfinance